refactor: cleanup and refactor sd chart
