+++ frontend/src/components/charts/supply-demand-chart.tsx
@@ -6,7 +6,6 @@ import {
     useMemo,
     useState,
     useRef,
-    useEffect,
     useCallback,
     memo,
     type ReactNode,
@@ -25,7 +24,13 @@ import {
 
 import { useAssetColorGetter } from "@/hooks/useAssetColorGetter";
 import { useMarketData } from "@/hooks/useCharts";
-import { getHashBasedChartColor } from "@/lib/charts/chart-utils";
+import { useRechartsChartArea } from "@/hooks/useRechartsChartArea";
+import {
+    getHashBasedChartColor,
+    interpolateAtX,
+    createSteppedCurve,
+    generateNiceTicks,
+} from "@/lib/charts/chart-utils";
 import { formatMoney, formatPower } from "@/lib/format-utils";
 
 export type BreakdownType = "supply" | "demand";
@@ -39,48 +44,6 @@ interface SupplyDemandChartProps {
     breakdownEnabled?: boolean;
     breakdownMode?: BreakdownMode;
     breakdownType?: BreakdownType;
-}
-
-interface CurvePoint {
-    quantity: number;
-    price: number;
-}
-
-/**
- * Performs linear interpolation on a curve to find the Y value at a given X
- * position. Returns null if X is outside the curve's range.
- */
-function interpolateAtX(curve: CurvePoint[], targetX: number): number | null {
-    if (curve.length === 0) return null;
-    if (targetX < 0) return null;
-
-    // Find the two points that bracket targetX
-    for (let i = 0; i < curve.length - 1; i++) {
-        const p1 = curve[i];
-        const p2 = curve[i + 1];
-
-        if (!p1 || !p2) continue;
-
-        // Check if targetX is between these two points
-        if (targetX >= p1.quantity && targetX <= p2.quantity) {
-            // Handle vertical segments (same quantity, different price)
-            if (p2.quantity === p1.quantity) {
-                return p2.price;
-            }
-
-            // Linear interpolation: y = y1 + (y2 - y1) * (x - x1) / (x2 - x1)
-            const ratio = (targetX - p1.quantity) / (p2.quantity - p1.quantity);
-            return p1.price + (p2.price - p1.price) * ratio;
-        }
-    }
-
-    // If targetX is beyond the last point, return the last price
-    const lastPoint = curve[curve.length - 1];
-    if (lastPoint && targetX >= lastPoint.quantity) {
-        return lastPoint.price;
-    }
-
-    return null;
 }
 
 interface CustomTooltipProps {
@@ -149,86 +112,6 @@ const OrderBlock = memo(function OrderBlock({
 });
 
 /**
- * Transforms order data into a stepped curve for Recharts. Creates points for a
- * ragged supply or demand curve from discrete orders.
- *
- * @param extensionPrice - Optional price to extend the curve vertically after
- *   the last order
- */
-function createSteppedCurve(
-    prices: number[],
-    cumulCapacities: number[],
-    extensionPrice?: number,
-): CurvePoint[] {
-    if (prices.length === 0) return [];
-
-    const points: CurvePoint[] = [];
-
-    // Start at origin
-    points.push({ quantity: 0, price: 0 });
-
-    for (let i = 0; i < prices.length; i++) {
-        const prevCumul = i === 0 ? 0 : (cumulCapacities[i - 1] ?? 0);
-        const currCumul = cumulCapacities[i] ?? 0;
-        const price = prices[i] ?? 0;
-
-        // Add vertical step (price change at same quantity)
-        points.push({ quantity: prevCumul, price });
-        // Add horizontal step (quantity change at same price)
-        points.push({ quantity: currCumul, price });
-    }
-
-    // Add vertical extension at the end if extensionPrice is provided
-    if (extensionPrice !== undefined && prices.length > 0) {
-        const lastQuantity = cumulCapacities[cumulCapacities.length - 1] ?? 0;
-        points.push({ quantity: lastQuantity, price: extensionPrice });
-    }
-
-    return points;
-}
-
-/**
- * Generates predictable tick values for an axis based on the range. Uses round
- * numbers to provide a consistent frame of reference.
- */
-function generateNiceTicks(
-    min: number,
-    max: number,
-    targetCount = 5,
-): number[] {
-    const range = max - min;
-    if (range === 0) return [min];
-
-    // Calculate order of magnitude
-    const roughStep = range / (targetCount - 1);
-    const magnitude = Math.pow(10, Math.floor(Math.log10(roughStep)));
-
-    // Try different nice step sizes (1, 2, 5 multiplied by magnitude)
-    const niceSteps = [1, 2, 5, 10].map((n) => n * magnitude);
-    const step =
-        niceSteps.find((s) => range / s <= targetCount) ??
-        niceSteps[niceSteps.length - 1] ??
-        magnitude;
-
-    // Generate ticks
-    const ticks: number[] = [];
-    const startTick = Math.floor(min / step) * step;
-
-    for (let tick = startTick; tick <= max; tick += step) {
-        if (tick >= min) {
-            ticks.push(tick);
-        }
-    }
-
-    // Ensure max is included
-    if (ticks[ticks.length - 1] !== max) {
-        ticks.push(max);
-    }
-
-    return ticks;
-}
-
-/**
  * Supply and demand chart showing market clearing with order blocks. Displays
  * ragged supply/demand curves from discrete orders, clearing point, and
  * individual order blocks.
@@ -261,14 +144,6 @@ function SupplyDemandChartInner({
 
     // State for currently hovered order block (only changes when block changes)
     const [hoveredBlockId, setHoveredBlockId] = useState<string | null>(null);
-
-    // State for measured chart plotting area (determined from Recharts internals)
-    const [chartArea, setChartArea] = useState<{
-        left: number;
-        top: number;
-        width: number;
-        height: number;
-    } | null>(null);
 
     // Ref to the container for coordinate calculations
     const containerRef = useRef<HTMLDivElement>(null);
@@ -442,56 +317,14 @@ function SupplyDemandChartInner({
         return scaleLinear().domain(priceDomain).range([1, 0]); // Inverted because Y coordinates increase downward
     }, [priceDomain]);
 
-    // Measure the actual chart plotting area after render by querying Recharts internals
-    // The CartesianGrid or clipPath gives us the exact plotting area coordinates
+    // Measure the actual chart plotting area from Recharts internals
+    // Re-measure when tick values change (affects layout)
     const maxPriceTick = priceTicks[priceTicks.length - 1];
     const maxQuantityTick = quantityTicks[quantityTicks.length - 1];
-    useEffect(() => {
-        if (!containerRef.current) return;
-
-        // Defer measurement to next animation frame to ensure Recharts has rendered
-        const rafId = requestAnimationFrame(() => {
-            if (!containerRef.current) return;
-
-            // Query for the CartesianGrid which defines the exact chart plotting area
-            // This is more reliable than measuring axes, which have labels in separate groups
-            const gridElement = containerRef.current.querySelector(
-                ".recharts-cartesian-grid",
-            ) as SVGGElement | null;
-
-            if (gridElement) {
-                try {
-                    // Get the bounding box of the grid - this IS the chart area
-                    const bbox = gridElement.getBBox();
-
-                    const newArea = {
-                        left: bbox.x,
-                        top: bbox.y,
-                        width: bbox.width,
-                        height: bbox.height,
-                    };
-
-                    // Only update if meaningfully changed
-                    setChartArea((prev) => {
-                        if (
-                            !prev ||
-                            Math.abs(prev.left - newArea.left) > 1 ||
-                            Math.abs(prev.top - newArea.top) > 1 ||
-                            Math.abs(prev.width - newArea.width) > 1 ||
-                            Math.abs(prev.height - newArea.height) > 1
-                        ) {
-                            return newArea;
-                        }
-                        return prev;
-                    });
-                } catch {
-                    // getBBox can fail in some edge cases, ignore
-                }
-            }
-        });
-
-        return () => cancelAnimationFrame(rafId);
-    }, [maxPriceTick, maxQuantityTick]); // Re-measure when ticks change
+    const chartArea = useRechartsChartArea(containerRef, [
+        maxPriceTick,
+        maxQuantityTick,
+    ]);
 
     // Mouse move handler to track cursor and interpolate values
     // Optimized to minimize state updates and re-renders
